set up main screen with placeholder
