+++ src/screens/TransactionLog.tsx
@@ -1,0 +1,23 @@
+import { StatusBar } from 'expo-status-bar';
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+function TransactionLog() {
+  return (
+    <View style={styles.container}>
+      <Text>This is a placeholder for the main screen of the app.</Text>
+      <StatusBar style="auto" />
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
+
+export default TransactionLog;
